requirements.txt
# Production Dependencies
# This package has no runtime dependencies by design - it uses only Python standard library

# Development Dependencies
# Install with: pip install -r requirements.txt

# Google ADK and Gemma LLM Integration Requirements
# Core ADK framework and LLM integration
google-adk
litellm
ollama
a2a-sdk
basic-open-agent-tools

# LLM and AI libraries
openai
transformers
torch
huggingface-hub

# Web scraping and API tools
requests
beautifulsoup4
selenium
aiohttp
httpx

# Data processing and analysis
pandas
numpy
python-dotenv>=1.0.0
pandas-stubs

# File and document processing
python-docx
openpyxl
markdown

# Development and project management tools
click
typer
rich
pydantic

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock

# ADK Evaluation Dependencies
tabulate>=0.9.0
rouge-score>=0.1.2

# Code Quality and Analysis
ruff>=0.1.0
mypy>=1.0.0
ast-decompiler
black
flake8

# Logging and monitoring
loguru
structlog

# Configuration management
pyyaml
toml
configparser

# Development Tools
pre-commit>=3.0.0

# Virtual environment and dependency management
virtualenv
pip-tools

# Optional: Jupyter for experimentation and learning
jupyter
ipykernel
notebook

# Optional: FastAPI for web interfaces (if using adk web)
fastapi
uvicorn

# Optional: Database support for advanced examples
sqlalchemy

# Optional: Vector databases for RAG examples
faiss-cpu